Changed timestamps to 32 bits; added auto-timestamp constructor.
git-svn-id: http://svn.psas.pdx.edu/svn/psas/trunk/java@199 6115d4fc-9e01-0410-8836-dad5f2c52705

+++ src/cansocket/TCPCanSocket.java
@@ -31,7 +31,7 @@ public class TCPCanSocket implements CanSocket
 	public CanMessage read() throws IOException
 	{
 		short id = din.readShort();
-		short timestamp = din.readShort();
+		int timestamp = din.readInt();
 		byte[] body = new byte[id & 0xF];
 		id = (short) ((id >> 5) & 0x7FF);
 		din.read(body);
@@ -41,7 +41,7 @@ public class TCPCanSocket implements CanSocket
 	public void write(CanMessage msg) throws IOException
 	{
 		dout.writeShort((msg.getId() & 0x7FF) << 5 | (msg.getBody().length & 0xF));
-		dout.writeShort(msg.getTimestamp());
+		dout.writeInt(msg.getTimestamp());
 		dout.write(msg.getBody());
 	}
